+++ src/main/java/me/pajic/simple_smithing_overhaul/items/WhetstoneItem.java
@@ -19,7 +19,7 @@ public class WhetstoneItem extends Item {
 
     @Override
     public boolean isEnabled(@NotNull FeatureFlagSet enabledFeatures) {
-        return SSO.CONFIG.whetstone.enableWhetstone.get();
+        return SSO.CONFIG.portableItemRepair.enableWhetstone.get();
     }
 
     @Override